fix(docker): prevent overwriting mounted config on container restart

+++ start.sh
@@ -1,8 +1,10 @@
 #/bin/sh
-if [ -f /app/public/conf/config_static.js ]; then
-  cp /app/public/conf/config_static.js /usr/share/nginx/html/conf/config.js
-elif [ ! -f /usr/share/nginx/html/conf/config.js ]; then
-  cp /app/public/conf/config.js /usr/share/nginx/html/conf
+if [ ! -f /usr/share/nginx/html/conf/config.js ]; then
+  if [ -f /app/public/conf/config_static.js ]; then
+    cp /app/public/conf/config_static.js /usr/share/nginx/html/conf/config.js
+  elif [ -f /app/public/conf/config.js ]; then
+    cp /app/public/conf/config.js /usr/share/nginx/html/conf
+  fi
 fi
 
 if [ $API_URL ]; then
